added weather test, minor reorder in tags test
